+++ packages/web/src/ShallotProvider/index.tsx
@@ -1,19 +1,15 @@
 import { ThemeTokens, ThemeVariants, makeTheme } from '@shallot-ui/core-theme'
-import { FunctionComponent, ReactNode, useMemo } from 'react'
+import { FunctionComponent, ReactNode } from 'react'
 import { ThemeProvider } from 'styled-components'
 
-import {
-  GlobalStyle,
-  GlobalStyleProps,
-  webVariants,
-} from '@shallot-ui/platform-react'
+import { GlobalStyle, webVariants } from '@shallot-ui/platform-react'
 
 type ShallotProviderProps = {
   children: ReactNode
   tokens?: ThemeTokens
   variants?: ThemeVariants
   includeGlobalStyle?: boolean
-} & GlobalStyleProps
+}
 
 export const ShallotProvider: FunctionComponent<ShallotProviderProps> = ({
   children,
@@ -26,7 +22,7 @@ export const ShallotProvider: FunctionComponent<ShallotProviderProps> = ({
 
   return (
     <ThemeProvider theme={theme}>
-      {includeGlobalStyle && <GlobalStyle {...rest} />}
+      {includeGlobalStyle && <GlobalStyle theme={theme} />}
       {children}
     </ThemeProvider>
   )