+++ components/Input/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @shallot-ui/component-input
 
+## 3.0.0-beta.7
+
+### Patch Changes
+
+- Fixes global styles
+- Updated dependencies
+  - @shallot-ui/core-mixins@2.0.0-beta.7
+  - @shallot-ui/core-theme@2.0.0-beta.7
+  - @shallot-ui/core-utils@2.0.0-beta.7
+
 ## 3.0.0-beta.6
 
 ### Patch Changes