+++ core/mixins/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @shallot-ui/core-mixins
 
+## 2.0.0-beta.7
+
+### Patch Changes
+
+- Fixes global styles
+- Updated dependencies
+  - @shallot-ui/core-theme@2.0.0-beta.7
+
 ## 2.0.0-beta.6
 
 ### Patch Changes