+++ core/utils/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @shallot-ui/core-utils
 
+## 2.0.0-beta.7
+
+### Patch Changes
+
+- Fixes global styles
+- Updated dependencies
+  - @shallot-ui/core-mixins@2.0.0-beta.7
+  - @shallot-ui/core-theme@2.0.0-beta.7
+
 ## 2.0.0-beta.6
 
 ### Patch Changes