+++ packages/native/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @shallot-ui/native
 
+## 3.0.0-beta.7
+
+### Patch Changes
+
+- Fixes global styles
+- Updated dependencies
+  - @shallot-ui/platform-react-native@2.0.0-beta.7
+  - @shallot-ui/component-checkbox@3.0.0-beta.7
+  - @shallot-ui/component-button@3.0.0-beta.7
+  - @shallot-ui/component-switch@3.0.0-beta.7
+  - @shallot-ui/component-input@3.0.0-beta.7
+  - @shallot-ui/component-tag@3.0.0-beta.7
+  - @shallot-ui/core-mixins@2.0.0-beta.7
+  - @shallot-ui/core-props@2.0.0-beta.7
+  - @shallot-ui/core-theme@2.0.0-beta.7
+  - @shallot-ui/core-utils@2.0.0-beta.7
+
 ## 3.0.0-beta.6
 
 ### Patch Changes